+++ packages/pages/src/exports/client.ts
@@ -1,3 +1,3 @@
 'use client'
 
-export { SlugInstruction } from '../admin/SlugInstruction.js'
+export { SlugInstruction } from '@sittari/payload-slug-field/client'